Add a new mode to EditTags, clean some css and add a few new classes
The new mode will be used for tag edit validation. It will show the latest change instead of '< keep >'. The code has been refactored to be easier to read and edit.
Some css propoerties were duplicated, so a new class 'mzk-input-colour-scheme 'was created to clean the code and make it easier to re-use.